+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## v0.9.0-rc17 [Unreleased]
 
-## Features
+### Features
 - [#2076](https://github.com/influxdb/influxdb/pull/2076): Seperate stdout and stderr output in init.d script
 
-## Bugfixes
+### Bugfixes
 - [#2084](https://github.com/influxdb/influxdb/pull/2084): Allowing leading underscores in identifiers.
+- [#2080](https://github.com/influxdb/influxdb/pull/2080): Graphite logs in seconds, not milliseconds.
 
 ## v0.9.0-rc16 [2015-03-24]